Camera Controls Basic
I couldn't really try the cameras in the main scene because it doesn't work, at least for me. But the script I wrote will work with every active Virtual Camera. So the moment you switch cameras between characters the controls will control the current active camera. Zoom with mousewheel and Q/E for rotation. Speed and limits for both can be set in the inspector.