Camera for Level 1 Test
So, I dragged the scripts/cameras over and changed some things. The camera works enough for testing that level. If more characters are added, then they need to be added to the target group of the main camera and get their own virtual cameras (with follow and aim target). I also get red error messages about something not being assigned in the mouseover script.